+++ src/search/tiles/tiles.cc
@@ -43,8 +43,8 @@ Tiles::Tiles(istream &in)
 	in >> width;
 
 	// Compute crazy Korf table.
-	ones.resize((int) pow(2, width * height - 1) + 1);
-	for (unsigned int i = 1; i <= pow(2, width * height - 1); i += 1) {
+	ones.resize((int) pow((double) 2.0, (int) (width * height - 1)) + 1);
+	for (unsigned int i = 1; i <= pow((double) 2.0, (int) (width * height - 1)); i += 1) {
 		unsigned int bits = 0;
 		unsigned int j = i;
